re-factor the code according to the comments done on the PR.
