+++ Src/DfT.DTRO/DAL/IApplicationDal.cs
@@ -7,6 +7,6 @@ namespace DfT.DTRO.DAL
         bool CheckApplicationNameDoesNotExist(string appName);
         ApplicationDetailsDto GetApplicationDetails(string appId);
         List<ApplicationListDto> GetApplicationList(string userId);
-        PaginatedResult<ApplicationListDto> GetPendingApplications(ApplicationRequest request);
+        PaginatedResult<ApplicationListDto> GetPendingApplications(PaginatedRequest request);
     }
 }
