--- Src/DfT.DTRO/Models/FIltering/ApplicationRequest.cs
@@ -1,7 +1,0 @@
-﻿namespace DfT.DTRO.Models.Filtering;
-
-public class ApplicationRequest : PaginatedRequest
-{
-    [DataType(DataType.EmailAddress)]
-    public string UserId { get; set; }
-}
