+++ Src/Dft.DTRO.Tests/GlobalUsings.cs
@@ -19,6 +19,7 @@ global using DfT.DTRO.Models.DtroHistory;
 global using DfT.DTRO.Models.DtroJson;
 global using DfT.DTRO.Models.Errors;
 global using DfT.DTRO.Models.Metrics;
+global using DfT.DTRO.Models.Pagination;
 global using DfT.DTRO.Models.Parameters;
 global using DfT.DTRO.Models.RuleTemplate;
 global using DfT.DTRO.Models.SchemaTemplate;
@@ -36,6 +37,7 @@ global using Microsoft.AspNetCore.Http;
 global using Microsoft.AspNetCore.Mvc;
 global using Microsoft.AspNetCore.Mvc.Testing;
 global using Microsoft.AspNetCore.TestHost;
+global using Microsoft.EntityFrameworkCore;
 global using Microsoft.Extensions.Configuration;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.Internal;
